sensors/requirements.txt
requests>=2.31.0
python-socketio[client]>=5.10.0
websocket-client>=1.6.0
spidev>=3.6; platform_machine == "aarch64"
rpi-lgpio>=0.6; platform_machine == "aarch64"
smbus2>=0.4.3; platform_machine == "aarch64"
